+++ tools/01_demo/STUFF/incident_robotshop-noi.sh
@@ -51,7 +51,7 @@ echo "   -----------------------------------------------------------------------
 
 export WAIOPS_NAMESPACE=$(oc get po -A|grep aiops-orchestrator-controller |awk '{print$1}')
 export EVTMGR_NAMESPACE=$(oc get po -A|grep noi-operator |awk '{print$1}')
-echo "       ✅ OK - AI Manager:    $WAIOPS_NAMESPACE"
+echo "       ✅ OK - CP4WAIOps:    $WAIOPS_NAMESPACE"
 echo "       ✅ OK - Event Manager: $EVTMGR_NAMESPACE"